crawlers/tiktok/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.6.1
python-dotenv==1.0.0
selenium==4.18.1
webdriver-manager==4.0.1 